Extract OrderSubscriptionTableViewCell to own file
